pyproject.toml
[project]
name = "python-be-genai-plgrnd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.10.6",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "ruff>=0.11.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
marks = [
    "integration: mark test as integration",
]

[tool.ruff]
indent-width = 4
line-length = 80

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # pyflakes
    "F",
    # flake8-bugbear
    "B",
    # isort
    "I",
]

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
docstring-code-format = true
